Projects/1-Q&A Chatbot/requirements.txt
streamlit
python-dotenv
langchain
langchain-groq
langchain-openai
langchain-community